+++ app/src/main/java/carnerero/agustin/cuentaappandroid/domain/database/accountusecase/UpdateAccountBalanceUseCase.kt
@@ -1,11 +1,12 @@
 package carnerero.agustin.cuentaappandroid.domain.database.accountusecase
 
 import carnerero.agustin.cuentaappandroid.data.repository.AccountRepository
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class UpdateAccountBalanceUseCase   @Inject constructor(private val repository: AccountRepository){
 
-    suspend operator fun invoke(accountId: Int, newBalance: Double) {
+    suspend operator fun invoke(accountId: Int, newBalance: BigDecimal) {
         return repository.updateAccountBalance(accountId, newBalance)
     }
 }
