+++ app/src/main/java/carnerero/agustin/cuentaappandroid/domain/database/entriesusecase/GetSumIncomesByMonthUseCase.kt
@@ -1,6 +1,7 @@
 package carnerero.agustin.cuentaappandroid.domain.database.entriesusecase
 
 import carnerero.agustin.cuentaappandroid.data.repository.EntryRepository
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class GetSumIncomesByMonthUseCase @Inject constructor(private val repository: EntryRepository) {
@@ -9,5 +10,5 @@ class GetSumIncomesByMonthUseCase @Inject constructor(private val repository: En
         accountId: Int,
         month: String,
         year:String
-    ): Double = repository.getSumOfIncomeEntriesForMonth(accountId, month, year)
+    ): BigDecimal= repository.getSumOfIncomeEntriesForMonth(accountId, month, year)
 }
