+++ app/src/main/java/carnerero/agustin/cuentaappandroid/domain/database/entriesusecase/GetSumOfExpensesByCategoryAndDateUseCase.kt
@@ -1,6 +1,7 @@
 package carnerero.agustin.cuentaappandroid.domain.database.entriesusecase
 
 import carnerero.agustin.cuentaappandroid.data.repository.EntryRepository
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class GetSumOfExpensesByCategoryAndDateUseCase @Inject constructor(private val repository: EntryRepository) {
@@ -9,5 +10,5 @@ class GetSumOfExpensesByCategoryAndDateUseCase @Inject constructor(private val r
         categoryId: Int,
         fromDate: String,
         toDate: String
-    ): Double = repository.getSumOfExpensesByCategoryAndDate(categoryId,fromDate,toDate)
+    ): BigDecimal = repository.getSumOfExpensesByCategoryAndDate(categoryId,fromDate,toDate)
 }
